Added all University hearings post, remove all other posts
