+++ .gitignore
@@ -11,8 +11,8 @@ screenshot.png
 readme_banner.svg
 History.markdown
 CODE_OF_CONDUCT.md
+/template_posts
 # /_json
-# Only Comment the Following 2
 # /_posts
