--- _posts/2023-11-30-This-Day-in-History.md
@@ -1,11 +1,0 @@
----
-layout: post
-title: "Today's News November 30, 2023"
-date: 2023-11-30
----
-
-| A Plane Fueled By Fat And Sugar Has Crossed The Atlantic Ocean | Flight To Honolulu On Route Of Commercial Line To Orient.  |
-|  |  |
-| Nicolás Rivero for The Washington Post | News Wire Article published in the Harrison Times |
-| November 30, 2023 | April 18, 1935 |
-| &nbsp;&nbsp;&nbsp;&nbsp;SAF, a broad category of jet fuel that creates fewer carbon emissions than standard kerosene blends. The fuel on this flight was made from waste fats and plant sugars and emits 70% less carbon than petroleum-based jet fuel, according to a press release. It landed at John F. Kennedy International Airport on Tuesday afternoon.<br>&nbsp;&nbsp;&nbsp;&nbsp;Experts say sustainable aviation fuels may one day play a big role in shrinking the aviation industry’s carbon footprint — even though its production is minuscule today. SAF accounts for about 0.1% of airlines’ current fuel consumption.<br>&nbsp;&nbsp;&nbsp;&nbsp;“SAF is a major aspect of the transition for aviation [to zero carbon emissions], and it’s especially critical this decade,” said Andrew Chen, principal for aviation decarbonization at the Rocky Mountain Institute, a clean-energy think tank. “But today, our big issue is we don’t make enough SAF.”<br>&nbsp;&nbsp;&nbsp;&nbsp;Virgin Atlantic’s 100% SAF flight is a one-time stunt, and the airline won’t regularly offer all-SAF flights. Standard jet engines aren ...<br><br>Read the full article at<br>[https://www.washingtonpost.com/climate-solutions/2023/11/28/virgin-atlantic-sustainable-flight-lands-saf/](https://www.washingtonpost.com/climate-solutions/2023/11/28/virgin-atlantic-sustainable-flight-lands-saf/) | &nbsp;&nbsp;&nbsp;&nbsp; <br>&nbsp;&nbsp;&nbsp;&nbsp;  Alameda, Calif., April 17.—Ex- factly on schedule, the big Clipper flying boat and its six-man crew left here Tuesday on a 2,400-mile flight to Honolulu to blaze the first leg of a commercial trans-<br>&nbsp;&nbsp;&nbsp;&nbsp;Pacifie airline to the Orient.<br>&nbsp;&nbsp;&nbsp;&nbsp;The Pan-American plane took off at 3:50:30 p. m. (6:50:30 eastern standard time), rose slowly from the water and headed toward its distant goal.<br>&nbsp;&nbsp;&nbsp;&nbsp;Four powerful motors, develop- ling 2,800 horsepower, whirled the big propellers as several thousand spectators eagerly watched the start of the pioneering hop.<br>&nbsp;&nbsp;&nbsp;&nbsp;The six men of the Clipper’s crew captained by Edwin C. Mu- sick, walked to the float and en- tered their craft. They wore natty blue uniforms and white caps.<br>&nbsp;&nbsp;&nbsp;&nbsp;Then Chief Pilot R. O. D. Sulli- van opened the throttles and the craft turned slowly and headed out into the bay.<br>&nbsp;&nbsp;&nbsp;&nbsp;Weather Reported Good<br>&nbsp;&nbsp;&nbsp;&nbsp;Spray foamed from the sleek black hull as the Clipper gathered speed. It raced along the surface for a mile and, after 30 seconds, lifted into the air.<br>&nbsp;&nbsp;&nbsp;&nbsp;Apparently flying slowly, — the Clipper nevertheless forced a dozen accompanying planes to race in keeping up with it as it passed over San Francisco and out the Golden Gate.<br>&nbsp;&nbsp;&nbsp;&nbsp;The two thin fingers of the Ciolden Gate bridge, under con- struction, formed a red-painted passageway for the plane as il lflew out over the blue Pacific. Weather conditions at sea were reported almost ideal.<br>&nbsp;&nbsp;&nbsp;&nbsp;With a cruising speed of between 150 and 160 miles an hour, the plane was expected to reach Hon- ‘Jolulu in between 16 and 18 hours.  |
